feature: add login system

+++ client/app/srcs/reservation/hooks/reservatio.hook.staff.ts
@@ -2,8 +2,10 @@ import { useCallback, useState } from 'react';
 import * as adminStaffApi from '../api/reservation.api.admin.staff';
 import { useStaffStore } from '../store/reservation.store.staff';
 import { ErrorType, useErrorStore } from '../store/reservation.store.error';
+import { useAuthStore } from '../store/reservation.store.auth';
 
 export const useStaff = () => {
+  const accessToken = useAuthStore((state) => state.accessToken);
   const staffs = useStaffStore((state) => state.staffs);
   const initStaff = useStaffStore((state) => state.initStaff);
   const addStaff = useStaffStore((state) => state.addStaff);
@@ -17,51 +19,53 @@ export const useStaff = () => {
     if (loading) return;
     setLoading(true);
     try {
-      const data = await adminStaffApi.getStaffs();
+      const data = await adminStaffApi.getStaffs({ accessToken });
       initStaff(data);
     } catch (err) {
       setErrorType(ErrorType.Unknown);
     } finally {
       setLoading(false);
     }
-  }, [initStaff, loading, setErrorType]);
+  }, [initStaff, loading, setErrorType, accessToken]);
 
   const createStaff = useCallback(async (name: string) => {
     if (loading) return;
     setLoading(true);
     try {
-      const staff = await adminStaffApi.postStaff({ name });
+      const staff = await adminStaffApi.postStaff({ name, accessToken });
       addStaff(staff);
     } catch (err) {
       setErrorType(ErrorType.Unknown);
     } finally {
       setLoading(false);
     }
-  }, [addStaff, loading, setErrorType]);
+  }, [addStaff, loading, setErrorType, accessToken]);
   const deleteStaffById = useCallback(async (id: string) => {
     if (loading) return;
     setLoading(true);
     try {
-      await adminStaffApi.deleteStaff(id);
+      await adminStaffApi.deleteStaff({ accessToken, id });
       deleteStaff(id);
     } catch (err) {
       setErrorType(ErrorType.Unknown);
     } finally {
       setLoading(false);
     }
-  }, [deleteStaff, loading, setErrorType]);
+  }, [deleteStaff, loading, setErrorType, accessToken]);
   const updateStaffById = useCallback(async (id: string, name: string) => {
     if (loading) return;
     setLoading(true);
     try {
-      await adminStaffApi.patchStaff(id, name);
+      await adminStaffApi.patchStaff({
+        accessToken, id, name,
+      });
       updateStaff(id, name);
     } catch (err) {
       setErrorType(ErrorType.Unknown);
     } finally {
       setLoading(false);
     }
-  }, [updateStaff, loading, setErrorType]);
+  }, [updateStaff, loading, setErrorType, accessToken]);
 
   return {
     staffs,
