+++ client/app/srcs/reservation/hooks/reservation.hook.nail.ts
@@ -2,8 +2,10 @@ import { useCallback, useState } from 'react';
 import { useNailStore } from '../store/reservation.store.nail';
 import * as adminNailApi from '../api/reservation.api.admin.nail';
 import { ErrorType, useErrorStore } from '../store/reservation.store.error';
+import { useAuthStore } from '../store/reservation.store.auth';
 
 export const useNail = () => {
+  const accessToken = useAuthStore((state) => state.accessToken);
   const nails = useNailStore((state) => state.nails);
   const initNails = useNailStore((state) => state.initNails);
   const addNail = useNailStore((state) => state.addNail);
@@ -17,53 +19,57 @@ export const useNail = () => {
     if (loading) return;
     setLoading(true);
     try {
-      const data = await adminNailApi.getNails();
+      const data = await adminNailApi.getNails({accessToken});
       initNails(data);
     } catch (err) {
       setErrorType(ErrorType.Unknown);
     } finally {
       setLoading(false);
     }
-  }, [initNails, loading, setErrorType]);
+  }, [initNails, loading, setErrorType, accessToken]);
 
   const createNail = useCallback(async (name: string, price: number, spendMinute: number) => {
     if (loading) return;
     setLoading(true);
     try {
-      const nail = await adminNailApi.postNail({ name, price, spendMinute });
+      const nail = await adminNailApi.postNail({
+        accessToken, name, price, spendMinute,
+      });
       addNail(nail);
     } catch (err) {
       setErrorType(ErrorType.Unknown);
     } finally {
       setLoading(false);
     }
-  }, [addNail, loading, setErrorType]);
+  }, [addNail, loading, setErrorType, accessToken]);
 
   const deleteNailById = useCallback(async (id: string) => {
     if (loading) return;
     setLoading(true);
     try {
-      await adminNailApi.deleteNail(id);
+      await adminNailApi.deleteNail({id, accessToken});
       deleteNail(id);
     } catch (err) {
       setErrorType(ErrorType.Unknown);
     } finally {
       setLoading(false);
     }
-  }, [deleteNail, loading, setErrorType]);
+  }, [deleteNail, loading, setErrorType, accessToken]);
 
   const updateNailById = useCallback(async (id: string, name: string, price: number) => {
     if (loading) return;
     setLoading(true);
     try {
-      await adminNailApi.patchNail(id, name, price);
+      await adminNailApi.patchNail({
+        id, name, price, accessToken,
+      });
       updateNail(id, name, price);
     } catch (err) {
       setErrorType(ErrorType.Unknown);
     } finally {
       setLoading(false);
     }
-  }, [updateNail, loading, setErrorType]);
+  }, [updateNail, loading, setErrorType, accessToken]);
 
   return {
     nails, fetchNails, createNail, deleteNailById, updateNail, updateNailById, loading,
